+++ hi-ml-cpath/src/health_cpath/utils/deepmil_utils.py
@@ -7,7 +7,7 @@ import param
 from torch import nn
 from pathlib import Path
 from typing import Optional, Tuple
-from health_ml.utils.checkpoint_utils import CheckpointDownloader
+from health_ml.utils.checkpoint_utils import CheckpointParser
 from health_cpath.models.encoders import (
     HistoSSLEncoder,
     ImageNetSimCLREncoder,
@@ -26,9 +26,6 @@ from health_ml.networks.layers.attention_layers import (
     TransformerPooling,
     TransformerPoolingBenchmark,
 )
-from health_ml.deep_learning_config import SRC_CHECKPOINT_FORMAT_DOC, CKPT_HELP_DOCS
-from health_ml.utils.checkpoint_handler import CheckpointHandler
-from health_ml.utils.common_utils import checkpoint_is_aml_run_id, checkpoint_is_local_file, checkpoint_is_url
 
 SSL_CHECKPOINT_DIRNAME = "ssl_checkpoint"
 
@@ -64,52 +61,8 @@ class EncoderParams(param.Parameterized):
     encoding_chunk_size: int = param.Integer(
         default=0, doc="If > 0 performs encoding in chunks, by enconding_chunk_size tiles " "per chunk"
     )
-    ssl_checkpoint: str = param.String(default="", doc=f"SSL checkpoint to use for the encoder. {CKPT_HELP_DOCS}")
-
-    @property
-    def ssl_checkpoint_is_url(self) -> bool:
-        return checkpoint_is_url(self.ssl_checkpoint)
-
-    @property
-    def ssl_checkpoint_is_local_file(self) -> bool:
-        return checkpoint_is_local_file(self.ssl_checkpoint)
-
-    @property
-    def ssl_checkpoint_is_aml_run_id(self) -> bool:
-        return checkpoint_is_aml_run_id(self.ssl_checkpoint)
-
-    @property
-    def is_valid_src_checkpoint(self) -> bool:
-        if self.ssl_checkpoint:
-            return self.ssl_checkpoint_is_local_file or self.ssl_checkpoint_is_url or self.ssl_checkpoint_is_aml_run_id
-        return True
-
-    def validate(self) -> None:
-        if not self.is_valid_src_checkpoint:
-            raise ValueError(f"Invalid ssl_checkpoint: {self.ssl_checkpoint}. Please provide a valid URL, local file "
-                             f"or azureml run id in the following format: {SRC_CHECKPOINT_FORMAT_DOC}")
-
-    def get_ssl_checkpoint_path(self, outputs_folder: Path) -> Path:
-        """
-        Get the path to the local weights if the ssl checkpoint is a local path, otherwise download the checkpoint
-        from the cloud and return the path to the downloaded checkpoint.
-        """
-
-        if self.ssl_checkpoint_is_local_file:
-            checkpoint_path = Path(self.ssl_checkpoint)
-        elif self.ssl_checkpoint_is_url:
-            download_folder = outputs_folder / SSL_CHECKPOINT_DIRNAME
-            download_folder.mkdir(exist_ok=True, parents=True)
-            checkpoint_path = CheckpointHandler.download_weights(self.ssl_checkpoint, download_folder)
-        elif self.ssl_checkpoint_is_aml_run_id:
-            downloader = CheckpointDownloader(run_id=self.ssl_checkpoint, download_dir=outputs_folder)
-            checkpoint_path = downloader.local_checkpoint_path
-        else:
-            raise ValueError("Unable to determine how to get the SSL checkpoint path.")
-
-        if checkpoint_path is None or not checkpoint_path.is_file():
-            raise FileNotFoundError(f"Could not find the weights file at {checkpoint_path}")
-        return checkpoint_path
+    ssl_checkpoint: CheckpointParser = param.ClassSelector(class_=CheckpointParser, default=CheckpointParser(),
+                                                           instantiate=False, doc=CheckpointParser.DOC)
 
     def get_encoder(self, outputs_folder: Optional[Path]) -> TileEncoder:
         """Given the current encoder parameters, returns the encoder object.
@@ -141,7 +94,7 @@ class EncoderParams(param.Parameterized):
         elif self.encoder_type == SSLEncoder.__name__:
             assert self.ssl_checkpoint and outputs_folder, "SSLEncoder requires ssl_checkpoint and outputs_folder"
             encoder = SSLEncoder(
-                pl_checkpoint_path=self.get_ssl_checkpoint_path(outputs_folder=outputs_folder),
+                pl_checkpoint_path=self.ssl_checkpoint.get_path(outputs_folder),
                 tile_size=self.tile_size,
                 n_channels=self.n_channels,
             )
