+++ hi-ml/src/health_ml/utils/checkpoint_utils.py
@@ -2,27 +2,32 @@
 #  Copyright (c) Microsoft Corporation. All rights reserved.
 #  Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 #  ------------------------------------------------------------------------------------------
+import re
+import os
+import uuid
+import torch
 import logging
 import tempfile
+import requests
+
 from pathlib import Path
 from typing import Optional
-
-import torch
+from urllib.parse import urlparse
 from azureml.core import Run, Workspace
 from health_azure import download_checkpoints_from_run_id, get_workspace
-
 from health_azure.utils import (RUN_CONTEXT, download_files_from_run_id, get_run_file_names, is_running_in_azure_ml)
-from health_ml.utils.common_utils import (AUTOSAVE_CHECKPOINT_CANDIDATES, DEFAULT_AML_CHECKPOINT_DIR)
+from health_ml.utils.common_utils import (AUTOSAVE_CHECKPOINT_CANDIDATES, DEFAULT_AML_CHECKPOINT_DIR, CHECKPOINT_SUFFIX)
 from health_ml.utils.type_annotations import PathOrString
+from health_azure.utils import is_global_rank_zero
+from health_ml.lightning_container import LightningContainer
 
-CHECKPOINT_SUFFIX = ".ckpt"
+
 # This is a constant that must match a filename defined in pytorch_lightning.ModelCheckpoint, but we don't want
 # to import that here.
-LAST_CHECKPOINT_FILE_NAME = "last"
-LAST_CHECKPOINT_FILE_NAME_WITH_SUFFIX = LAST_CHECKPOINT_FILE_NAME + CHECKPOINT_SUFFIX
+LAST_CHECKPOINT_FILE_NAME = f"last{CHECKPOINT_SUFFIX}"
 LEGACY_RECOVERY_CHECKPOINT_FILE_NAME = "recovery"
 MODEL_INFERENCE_JSON_FILE_NAME = "model_inference_config.json"
-MODEL_WEIGHTS_DIR_NAME = "trained_models"
+MODEL_WEIGHTS_DIR_NAME = "pretrained_models"
 
 
 def get_best_checkpoint_path(path: Path) -> Path:
@@ -31,7 +36,7 @@ def get_best_checkpoint_path(path: Path) -> Path:
 
     :param path to checkpoint folder
     """
-    return path / LAST_CHECKPOINT_FILE_NAME_WITH_SUFFIX
+    return path / LAST_CHECKPOINT_FILE_NAME
 
 
 def download_folder_from_run_to_temp_folder(folder: str,
@@ -120,7 +125,7 @@ def find_recovery_checkpoint(path: Path) -> Optional[Path]:
         logging.warning(f"Found these legacy checkpoint files: {legacy_recovery_checkpoints}")
         raise ValueError("The legacy recovery checkpoint setup is no longer supported. As a workaround, you can take "
                          f"one of the legacy checkpoints and upload as '{AUTOSAVE_CHECKPOINT_CANDIDATES[0]}'")
-    candidates = [*AUTOSAVE_CHECKPOINT_CANDIDATES, LAST_CHECKPOINT_FILE_NAME_WITH_SUFFIX]
+    candidates = [*AUTOSAVE_CHECKPOINT_CANDIDATES, LAST_CHECKPOINT_FILE_NAME]
     highest_epoch: Optional[int] = None
     file_with_highest_epoch: Optional[Path] = None
     for f in candidates:
@@ -147,10 +152,10 @@ def cleanup_checkpoints(ckpt_folder: Path) -> None:
     if len(files_in_checkpoint_folder) == 0:
         return
     logging.info(f"Files in checkpoint folder: {' '.join(files_in_checkpoint_folder)}")
-    last_ckpt = ckpt_folder / LAST_CHECKPOINT_FILE_NAME_WITH_SUFFIX
+    last_ckpt = ckpt_folder / LAST_CHECKPOINT_FILE_NAME
     all_files = f"Existing files: {' '.join(p.name for p in ckpt_folder.glob('*'))}"
     if not last_ckpt.is_file():
-        raise FileNotFoundError(f"Checkpoint file {LAST_CHECKPOINT_FILE_NAME_WITH_SUFFIX} not found. {all_files}")
+        raise FileNotFoundError(f"Checkpoint file {LAST_CHECKPOINT_FILE_NAME} not found. {all_files}")
     # Training is finished now. To save storage, remove the autosave checkpoint which is now obsolete.
     # Lightning does not overwrite checkpoints in-place. Rather, it writes "autosave.ckpt",
     # then "autosave-1.ckpt" and deletes "autosave.ckpt", then "autosave.ckpt" and deletes "autosave-v1.ckpt"
@@ -160,7 +165,7 @@ def cleanup_checkpoints(ckpt_folder: Path) -> None:
             autosave.unlink()
 
 
-class CheckpointDownloader:
+class AMLCheckpointDownloader:
     def __init__(
         self,
         run_id: str,
@@ -183,7 +188,6 @@ class CheckpointDownloader:
         self.remote_checkpoint_dir = (
             remote_checkpoint_dir or self.extract_remote_checkpoint_dir_from_checkpoint_filename()
         )
-        self.download_checkpoint_if_necessary()
 
     def extract_checkpoint_filename_from_run_id(self) -> str:
         """
@@ -192,7 +196,7 @@ class CheckpointDownloader:
         """
         run_id_split = self.run_id.split(":")
         self.run_id = run_id_split[0]
-        return run_id_split[-1] if len(run_id_split) > 1 else LAST_CHECKPOINT_FILE_NAME_WITH_SUFFIX
+        return run_id_split[-1] if len(run_id_split) > 1 else LAST_CHECKPOINT_FILE_NAME
 
     def extract_remote_checkpoint_dir_from_checkpoint_filename(self) -> Path:
         """
@@ -232,3 +236,168 @@ class CheckpointDownloader:
                 self.run_id, str(self.remote_checkpoint_path), self.local_checkpoint_dir, aml_workspace=workspace
             )
             assert self.local_checkpoint_path.exists(), f"Couln't download checkpoint from run {self.run_id}."
+
+
+class CheckpointHandler:
+    """
+    This class handles which checkpoints are used to initialize the model during train or test time
+    """
+
+    def __init__(self, container: LightningContainer, project_root: Path, run_context: Optional[Run] = None):
+        self.container = container
+        self.project_root = project_root
+        self.run_context = run_context
+        self.trained_weights_path: Optional[Path] = None
+        self.has_continued_training = False
+
+    def download_recovery_checkpoints_or_weights(self) -> None:
+        """
+        Download checkpoints from a run recovery object or from a given checkpoint. Set the checkpoints path based on
+        the checkpoint_url, local_checkpoint or checkpoint from an azureml run id.
+        This is called at the start of training.
+        """
+
+        if self.container.src_checkpoint:
+            self.trained_weights_path = self.get_local_checkpoints_path_or_download()
+            self.container.trained_weights_path = self.trained_weights_path
+
+    def additional_training_done(self) -> None:
+        """
+        Lets the handler know that training was done in this run.
+        """
+        self.has_continued_training = True
+
+    def get_recovery_or_checkpoint_path_train(self) -> Optional[Path]:
+        """
+        Decides the checkpoint path to use for the current training run. Looks for the latest checkpoint in the
+        checkpoint folder. If run_recovery is provided, the checkpoints will have been downloaded to this folder
+        prior to calling this function. Else, if the run gets pre-empted and automatically restarted in AML,
+        the latest checkpoint will be present in this folder too.
+
+        :return: Constructed checkpoint path to recover from.
+        """
+        if is_global_rank_zero():
+            checkpoints = list(self.container.checkpoint_folder.rglob("*"))
+            logging.info(f"Available checkpoints: {len(checkpoints)}")
+            for f in checkpoints:
+                logging.info(f)
+        return find_recovery_checkpoint_on_disk_or_cloud(self.container.checkpoint_folder)
+
+    def get_checkpoint_to_test(self) -> Path:
+        """
+        Find the model checkpoint that should be used for inference. If the model
+        has been training, get the best checkpoint as defined by the container.
+        If the model was not trained in this run, then return the checkpoint from the trained_weights_path.
+        """
+        if self.has_continued_training:
+            # If model was trained, look for the best checkpoint
+            checkpoint_from_current_run = self.container.get_checkpoint_to_test()
+            if checkpoint_from_current_run.is_file():
+                logging.info(f"Using checkpoint from current run: {checkpoint_from_current_run}")
+                return checkpoint_from_current_run
+            else:
+                raise FileNotFoundError(f"Checkpoint file does not exist: {checkpoint_from_current_run}")
+        elif self.trained_weights_path:
+            # Model was not trained, check if there is a local weight path.
+            logging.info(f"Using pre-trained weights from {self.trained_weights_path}")
+            return self.trained_weights_path
+        raise ValueError("Unable to determine which checkpoint should be used for testing.")
+
+    def get_local_checkpoints_path_or_download(self) -> Path:
+        """
+        Get the path to the local weights to use or download them.
+        """
+        return self.container.src_checkpoint.get_path(self.container.checkpoint_folder)
+
+
+class CheckpointParser:
+    AML_RUN_ID_FORMAT = (f"<AzureML_run_id>:<optional/custom/path/to/checkpoints/><filename{CHECKPOINT_SUFFIX}>"
+                         f"If no custom path is provided (e.g., <AzureML_run_id>:<filename{CHECKPOINT_SUFFIX}>)"
+                         "the checkpoint will be downloaded from the default checkpoint folder "
+                         f"(e.g., '{DEFAULT_AML_CHECKPOINT_DIR}') If no filename is provided, "
+                         "(e.g., `src_checkpoint=<AzureML_run_id>`) the latest checkpoint "
+                         f"({LAST_CHECKPOINT_FILE_NAME}) will be used to initialize the model.")
+    INFO_MESSAGE = ("Please provide a valid checkpoint path, URL or AzureML run ID. For custom checkpoint paths "
+                    f"within an azureml run, provide a checkpoint in the format {AML_RUN_ID_FORMAT}.")
+    DOC = ("We currently support three types of checkpoints: "
+           "    a. A local checkpoint folder that contains a checkpoint file."
+           "    b. A URL to a remote checkpoint to be downloaded."
+           "    c. A previous azureml run id where the checkpoint is supposed to be "
+           "       saved ('outputs/checkpoints/' folder by default.)"
+           f"For the latter case 'c' : src_checkpoint should be in the format of {AML_RUN_ID_FORMAT}")
+
+    def __init__(self, checkpoint: str = "") -> None:
+        self.checkpoint = checkpoint
+        self.validate()
+
+    @property
+    def is_url(self) -> bool:
+        try:
+            result = urlparse(self.checkpoint)
+            return all([result.scheme, result.netloc])
+        except ValueError:
+            return False
+
+    @property
+    def is_local_file(self) -> bool:
+        return Path(self.checkpoint).is_file()
+
+    @property
+    def is_aml_run_id(self) -> bool:
+        match = re.match(r"[_\w-]*$", self.checkpoint.split(":")[0])
+        return match is not None and not self.is_url and not self.is_local_file
+
+    @property
+    def is_valid(self) -> bool:
+        if self.checkpoint:
+            return self.is_local_file or self.is_url or self.is_aml_run_id
+        return True
+
+    def validate(self) -> None:
+        if not self.is_valid:
+            raise ValueError(f"Invalid checkpoint '{self.checkpoint}'. {self.INFO_MESSAGE}")
+
+    @staticmethod
+    def download_weights(url: str, download_folder: Path) -> Path:
+        """
+        Download a checkpoint from checkpoint_url to the modelweights directory. The file name is determined from
+        from the file name in the URL. If that can't be determined, use a random file name.
+
+        :param url: The URL from which the weights should be downloaded.
+        :param download_folder: The target folder for the download.
+        :return: A path to the downloaded file.
+        """
+        # assign the same filename as in the download url if possible, so that we can check for duplicates
+        # If that fails, map to a random uuid
+        file_name = os.path.basename(urlparse(url).path) or str(uuid.uuid4().hex)
+        checkpoint_path = download_folder / file_name
+        # only download if hasn't already been downloaded
+        if checkpoint_path.is_file():
+            logging.info(f"File already exists, skipping download: {checkpoint_path}")
+        else:
+            logging.info(f"Downloading weights from URL {url}")
+
+            response = requests.get(url, stream=True)
+            response.raise_for_status()
+            with open(checkpoint_path, "wb") as file:
+                for chunk in response.iter_content(chunk_size=1024):
+                    file.write(chunk)
+        return checkpoint_path
+
+    def get_path(self, checkpoints_folder: Path) -> Path:
+        if self.is_local_file:
+            checkpoint_path = Path(self.checkpoint)
+        elif self.is_url:
+            download_folder = checkpoints_folder / MODEL_WEIGHTS_DIR_NAME
+            download_folder.mkdir(exist_ok=True, parents=True)
+            checkpoint_path = self.download_weights(url=self.checkpoint, download_folder=download_folder)
+        elif self.is_aml_run_id:
+            downloader = AMLCheckpointDownloader(run_id=self.checkpoint, download_dir=checkpoints_folder)
+            downloader.download_checkpoint_if_necessary()
+            checkpoint_path = downloader.local_checkpoint_path
+        else:
+            raise ValueError("Unable to determine how to get the checkpoint path.")
+
+        if checkpoint_path is None or not checkpoint_path.is_file():
+            raise FileNotFoundError(f"Could not find the weights file at {checkpoint_path}")
+        return checkpoint_path
