+++ hi-ml/testhiml/testhiml/test_deep_learning_config.py
@@ -18,7 +18,7 @@ from testhiml.utils.fixed_paths_for_tests import full_test_data_path, mock_run_i
 
 def _test_invalid_src_checkpoint_workflow_params(src_checkpoint: str) -> None:
     with pytest.raises(ValueError, match=r"Invalid src_checkpoint:"):
-        WorkflowParams(local_datasets=Path("foo"), src_checkpoint=src_checkpoint).validate()
+        WorkflowParams(local_datasets=Path("foo"), src_checkpoint=src_checkpoint)
 
 
 def test_validate_workflow_params_src_checkpoint() -> None:
@@ -29,9 +29,9 @@ def test_validate_workflow_params_src_checkpoint() -> None:
 
     # The following should be okay
     full_file_path = full_test_data_path(suffix="hello_world_checkpoint.ckpt")
-    WorkflowParams(local_dataset=Path("foo"), src_checkpoint=str(full_file_path)).validate()
+    WorkflowParams(local_dataset=Path("foo"), src_checkpoint=str(full_file_path))
     run_id = mock_run_id(id=0)
-    WorkflowParams(local_dataset=Path("foo"), src_checkpoint=run_id).validate()
+    WorkflowParams(local_dataset=Path("foo"), src_checkpoint=run_id)
 
 
 def test_validate_workflow_params_for_inference_only() -> None:
